ingestion/requirements.txt
kafka-python
websocket-client
pandas
requests
s3fs
pyarrow